requirements.txt
requests==2.31.0
Pillow==9.5.0
python-pptx==0.6.21